Historical EIA Data Integration

+++ workflow/scripts/retrieve_historical_load_data.py
@@ -14,16 +14,20 @@ import re
 from pathlib import Path
 import warnings
 
-def download_historical_load_data(url, pattern):
-    pattern = r'EBA\..*-ALL\.D\.H'  # Define the header filter pattern
+
+def download_historical_load_data(url, year):
+    # pattern = r'EBA\..*-ALL\.D\.H'  # Define the header filter pattern
+    pattern = r'EBA\.(.*?)-ALL\.D\.H'
+
     response = requests.get(url)
-    if response.status_code == 200:      # Check if the request was successful
+    if response.status_code == 200:  # Check if the request was successful
         buffer = BytesIO(response.content)
         with gzip.open(buffer, 'rt') as file:
             df = pd.read_csv(file)
             filtered_columns = [col for col in df.columns if re.match(pattern, col)]
             filtered_df = df[filtered_columns]
-            # import pdb; pdb.set_trace() 
+            updated_columns = [re.search(pattern, col).group(1) for col in filtered_columns]
+            filtered_df.columns = updated_columns
             filtered_df.insert(0,"timestamp", "")
             filtered_df.iloc[:,0] = pd.to_datetime(df['Unnamed: 0'])
             df = filtered_df.set_index('timestamp')
@@ -45,16 +49,20 @@ if __name__ == "__main__":
     rootpath = "./"
     PATH_DOWNLOAD = Path(f"{rootpath}/resources/eia")
     PATH_DOWNLOAD.mkdir(parents=True, exist_ok=True)
-    i=0
-    for year in range(2015, 2024):
-        if year >= 2019:
-            df = download_historical_load_data(url_2018_present, year)
-        elif year == 2018:
-            df_1 = download_historical_load_data(url_2015_2018, year)
-            df_2 = download_historical_load_data(url_2018_present, year)
-            df = pd.concat([df_1.iloc[:-1, :], df_2])  # one duplicate row removed
-        else:
-            df = download_historical_load_data(url_2015_2018, year)
-        df.to_csv(os.path.join(snakemake.output[i]))
-        print('saving to ', os.path.join(snakemake.output[i]))
-        i += 1
+    i = 0
+
+    if os.path.isfile(os.path.join(PATH_DOWNLOAD, snakemake.output[6])):
+        logger.info("EIA Data bundle already downloaded.")
+    else:
+        for year in range(2015, 2024):
+            if year >= 2019:
+                df = download_historical_load_data(url_2018_present, year)
+            elif year == 2018:
+                df_1 = download_historical_load_data(url_2015_2018, year)
+                df_2 = download_historical_load_data(url_2018_present, year)
+                df = pd.concat([df_1.iloc[:-1, :], df_2])  # one duplicate row removed
+            else:
+                df = download_historical_load_data(url_2015_2018, year)
+            df.to_csv(os.path.join(snakemake.output[i]))
+            logger.info('saving to ', os.path.join(snakemake.output[i]))
+            i += 1
